limit visibilites to crate. only expose necessasry items to main.
