+++ src/lib.rs
@@ -1,6 +1,6 @@
-pub mod board;
-pub mod chain;
 pub mod game;
-pub mod game_move;
-pub mod stone;
 pub mod ui;
+pub(crate) mod board;
+pub(crate) mod chain;
+pub(crate) mod game_move;
+pub(crate) mod stone;
