+++ src/ui/text_ui.rs
@@ -1,22 +1,7 @@
-//! The ui module should contain all user interaction code. This allows your other code to focus on
-//! game logic rather than handling inputs/outputs.
+use super::*;
 use crate::board::Board;
 use anyhow::{bail, Context, Result};
 use std::io::{BufRead, BufReader, BufWriter, Read, Write};
-
-#[derive(Debug, PartialEq, Eq)]
-pub enum UserAction {
-    Move(usize, usize),
-    Quit,
-}
-
-pub trait UserInterface {
-    /// User input, which will be passed to the controller
-    fn input(&mut self) -> Result<UserAction>;
-
-    /// View the model
-    fn view(&mut self, board: &Board) -> Result<()>;
-}
 
 #[derive(Debug)]
 pub struct TextUi<R: Read, W: Write> {
